Added chat session naming
In order to log session conversations and review them in a logical manner
we need a way to name the sessions so all conversations can be grouped together.

Upon application start each user will create a session name based on the pattern
(<my_name>_<other_party_name>_MMDDYYYY_HHMMSS). The name created by the party in
the chat who issues the first message in the converstion will the the name of
the session.

This value is send as a MessageAttribute whenever a message is published. This
will allow the lambda function that handles logging to pass along the session
identifier.